requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Rate Limiting
slowapi==0.1.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Pin to 4.x for passlib compatibility (5.x has breaking changes)
authlib==1.3.0  # OAuth client library for social login
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (for external APIs if needed)
httpx==0.26.0

# Elasticsearch (Full-text Search)
elasticsearch[async]==8.11.0

# GraphQL
strawberry-graphql[fastapi]==0.217.0

# Scientific Computing (for recommendations)
numpy==1.26.0

# Utilities
python-dateutil==2.8.2
